refactor: create index for Icons, implement Tailwind variants, extract and create InputLabel and InputWrapper as components
