+++ app/components/Icons/Confetti.tsx
@@ -1,58 +1,48 @@
-export default function Confetti(props: React.SVGProps<SVGSVGElement>) {
+export function Confetti(props: React.SVGProps<SVGSVGElement>) {
   return (
     <svg
       width="62"
       height="62"
       viewBox="0 0 62 62"
-      fill="none"
       xmlns="http://www.w3.org/2000/svg"
       {...props}
     >
       <path
         d="M28.2265 6.09693C29.1394 6.65505 29.4271 7.84761 28.869 8.76058C28.464 9.42307 28.5655 10.2768 29.1145 10.8258L29.3674 11.0787C30.8881 12.5994 31.4488 14.8334 30.8266 16.892C30.517 17.9163 29.4356 18.4956 28.4113 18.186C27.3871 17.8764 26.8077 16.795 27.1173 15.7708C27.3263 15.0795 27.138 14.3294 26.6274 13.8187L26.3745 13.5659C24.5596 11.751 24.2241 8.92924 25.5628 6.73943C26.121 5.82646 27.3135 5.53881 28.2265 6.09693Z"
-        fill="#F87171"
         className={props.className}
       />
       <path
         d="M45.6969 12.1975C46.7462 12.4073 47.4267 13.428 47.2168 14.4773L46.8449 16.3371C46.3329 18.8967 44.4888 20.9858 42.0124 21.8113C40.8553 22.197 39.9935 23.1731 39.7543 24.3692L39.3824 26.229C39.1725 27.2783 38.1518 27.9588 37.1025 27.7489C36.0532 27.5391 35.3728 26.5183 35.5826 25.4691L35.9546 23.6092C36.4665 21.0496 38.3107 18.9606 40.787 18.1351C41.9442 17.7494 42.8059 16.7732 43.0451 15.5772L43.4171 13.7174C43.6269 12.6681 44.6477 11.9876 45.6969 12.1975Z"
-        fill="#F87171"
         className={props.className}
       />
       <path
         fill-rule="evenodd"
         clip-rule="evenodd"
         d="M30.2726 22.7668C28.1578 20.6519 26.4151 18.9091 24.9004 17.7861C23.3355 16.6259 21.5964 15.8083 19.5935 16.2811C17.5906 16.7539 16.4007 18.263 15.5199 20.0006C14.6674 21.6824 13.8881 24.0205 12.9423 26.8579L8.47307 40.2658C7.32777 43.7014 6.39837 46.4894 6.01379 48.6513C5.62574 50.8327 5.6649 53.0462 7.30816 54.6895C8.95143 56.3327 11.165 56.3719 13.3463 55.9838C15.5081 55.5993 18.296 54.6699 21.7315 53.5247L35.1397 49.0553C37.977 48.1096 40.3153 47.3303 41.9971 46.4777C43.7347 45.5969 45.2437 44.407 45.7165 42.4041C46.1894 40.4013 45.3718 38.6621 44.2115 37.0972C43.0885 35.5825 41.3458 33.8399 39.2308 31.725L30.2726 22.7668ZM22.4599 20.802C22.5032 20.8331 22.5474 20.8654 22.5926 20.8989C23.8483 21.8299 25.3888 23.3631 27.6471 25.6214L29.6099 27.5842C29.571 27.6858 29.5279 27.8006 29.4815 27.9277C29.2825 28.4732 29.0208 29.2504 28.76 30.201C28.241 32.0923 27.7116 34.7203 27.7116 37.6014C27.7116 40.3452 28.2433 42.9739 28.7574 44.8851C28.9715 45.681 29.1859 46.3647 29.3648 46.8956L22.5586 49.1644C22.5305 49.0371 22.4995 48.895 22.4658 48.7395C22.3084 48.0118 22.0954 46.9935 21.8731 45.833C21.4242 43.4898 20.954 40.6495 20.8101 38.4539C20.5723 34.8274 21.0368 30.095 21.5861 26.1804C21.8581 24.2419 22.1458 22.5382 22.3654 21.3198C22.3985 21.1358 22.4301 20.9629 22.4599 20.802ZM17.9651 24.1633C17.5561 25.2751 17.1071 26.6176 16.5673 28.237L12.2027 41.3306C10.9912 44.9652 10.1576 47.482 9.8289 49.33C9.49766 51.192 9.80594 51.7072 10.0482 51.9494C10.2905 52.1917 10.8057 52.5 12.6677 52.1687C14.1929 51.8974 16.1737 51.2822 18.8625 50.3945C18.8163 50.189 18.7534 49.9055 18.6784 49.5586C18.5161 48.8083 18.2966 47.7592 18.0673 46.562C17.6129 44.1901 17.1038 41.1545 16.9434 38.7075C16.6782 34.6634 17.1929 29.6027 17.7487 25.6419C17.8203 25.1317 17.8929 24.6374 17.9651 24.1633ZM33.041 45.6702L33.7606 45.4304C36.7904 44.4204 38.8507 43.7283 40.245 43.0214C41.6139 42.3275 41.8653 41.8521 41.9452 41.5138C42.0251 41.1755 42.0128 40.6379 41.0987 39.4051C40.1677 38.1493 38.6345 36.6088 36.3762 34.3505L32.665 30.6394C32.6096 30.8248 32.5533 31.0208 32.4969 31.2264C32.0365 32.9042 31.5866 35.1728 31.5866 37.6014C31.5866 39.1995 31.8065 40.7917 32.0983 42.1968C32.2228 42.7961 32.3603 43.3614 32.4994 43.8785C32.6916 44.5931 32.8836 45.2037 33.041 45.6702Z"
-        fill="#F87171"
         className={props.className}
       />
       <path
         d="M52.5204 32.8855C53.2835 32.2089 54.3702 32.0373 55.3046 32.4459L56.0576 32.7751C57.038 33.2038 58.1803 32.7565 58.609 31.7761C59.0377 30.7956 58.5904 29.6533 57.61 29.2246L56.857 28.8954C54.5387 27.8817 51.8426 28.3075 49.9495 29.9862C49.1069 30.7334 47.8806 30.8576 46.9053 30.2945L46.3554 29.9771C45.4288 29.442 44.2438 29.7595 43.7088 30.6862C43.1737 31.6129 43.4913 32.7979 44.4179 33.3329L44.9678 33.6504C47.3876 35.0474 50.4299 34.7393 52.5204 32.8855Z"
-        fill="#F87171"
         className={props.className}
       />
       <path
         d="M36.112 10.4829C35.8117 10.5781 35.552 10.8378 35.0326 11.3572C34.5132 11.8766 34.2535 12.1363 34.1583 12.4365C34.0778 12.6905 34.0778 12.9632 34.1583 13.2172C34.2535 13.5175 34.5132 13.7772 35.0326 14.2966C35.552 14.816 35.8117 15.0757 36.112 15.1709C36.366 15.2514 36.6386 15.2514 36.8926 15.1709C37.1929 15.0757 37.4526 14.816 37.972 14.2966C38.4914 13.7772 38.7511 13.5175 38.8463 13.2172C38.9268 12.9632 38.9268 12.6905 38.8463 12.4365C38.7511 12.1363 38.4914 11.8766 37.972 11.3572C37.4526 10.8378 37.1929 10.5781 36.8926 10.4829C36.6386 10.4024 36.366 10.4024 36.112 10.4829Z"
-        fill="#F87171"
         className={props.className}
       />
       <path
         d="M50.293 19.2929C50.9269 18.659 51.2439 18.342 51.6015 18.2043C52.0199 18.043 52.4834 18.043 52.9018 18.2043C53.2594 18.342 53.5763 18.659 54.2103 19.2929C54.8442 19.9268 55.1611 20.2438 55.2989 20.6014C55.4601 21.0198 55.4601 21.4832 55.2989 21.9017C55.1611 22.2593 54.8442 22.5762 54.2103 23.2101C53.5763 23.8441 53.2594 24.161 52.9018 24.2988C52.4834 24.46 52.0199 24.46 51.6015 24.2988C51.2439 24.161 50.9269 23.8441 50.293 23.2101C49.6591 22.5762 49.3421 22.2593 49.2044 21.9017C49.0431 21.4832 49.0431 21.0198 49.2044 20.6014C49.3421 20.2438 49.6591 19.9268 50.293 19.2929Z"
-        fill="#F87171"
         className={props.className}
       />
       <path
         d="M19.8554 10.1804C19.3143 9.63926 18.4369 9.63926 17.8958 10.1804C17.3547 10.7215 17.3547 11.5989 17.8958 12.14C18.4369 12.6811 19.3143 12.6811 19.8554 12.14C20.3965 11.5989 20.3965 10.7215 19.8554 10.1804Z"
-        fill="#F87171"
         className={props.className}
       />
       <path
         d="M49.2339 39.5595C49.775 39.0184 50.6523 39.0184 51.1935 39.5595C51.7346 40.1006 51.7346 40.978 51.1935 41.5191C50.6523 42.0602 49.775 42.0602 49.2339 41.5191C48.6927 40.978 48.6927 40.1006 49.2339 39.5595Z"
-        fill="#F87171"
         className={props.className}
       />
       <path
         d="M47.168 25.1654C46.6269 24.6243 45.7496 24.6243 45.2084 25.1654C44.6673 25.7065 44.6673 26.5839 45.2084 27.125C45.7496 27.6661 46.6269 27.6661 47.168 27.125C47.7092 26.5839 47.7092 25.7065 47.168 25.1654Z"
-        fill="#F87171"
         className={props.className}
       />
     </svg>
