+++ app/components/Icons/Letter.tsx
@@ -1,4 +1,4 @@
-export default function Letter(props: React.SVGProps<SVGSVGElement>) {
+export function Letter(props: React.SVGProps<SVGSVGElement>) {
   return (
     <svg
       width="20"
