+++ app/components/Icons/index.ts
@@ -1,0 +1,6 @@
+export { User } from "./User";
+export { Letter } from "./Letter";
+export { Lock } from "./Lock";
+export { LoadingSpinner } from "./LoadingSpinner";
+export { Confetti } from "./Confetti";
+export { EyeClosed } from "./EyeClosed";
