+++ package.json
@@ -12,7 +12,8 @@
     "@next/font": "^14.2.14",
     "next": "^14.2.15",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "tailwind-variants": "^0.2.1"
   },
   "devDependencies": {
     "@types/node": "^20",
