+++ tailwind.config.ts
@@ -13,6 +13,7 @@ const colors = {
     900: "#1C202E",
   },
   complementary: {
+    400: "#F87171",
     500: "#EF4444",
     600: "#DC2626",
     700: "#B91C1C",
